Created a basic profile page
Creating a new now creates an entry in the Team table in addition to the entry created in the User table. When loaded, the profile page displays the team name, owner, current record, and current roster (this is currently empty because when you create a new team, the roster is empty).